app/requirements.txt
starlette==0.46.1
uvicorn[standard]==0.34.0
uvloop==0.21.0
httpx==0.28.1
pyyaml==6.0.2
